+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useDefaultBlocksLogic/useDefaultBlocksLogic.tsx
@@ -50,6 +50,7 @@ import { useLocation } from 'react-router-dom';
 import { omitPropsFromObjectWhitelist } from '@/common/utils/omit-props-from-object-whitelist/omit-props-from-object-whitelist';
 import { useObjectEntriesBlock } from '@/lib/blocks/hooks/useObjectEntriesBlock/useObjectEntriesBlock';
 import { useAmlBlock } from '@/lib/blocks/components/AmlBlock/hooks/useAmlBlock/useAmlBlock';
+import { associatedCompanyToWorkflowAdapter } from '@/lib/blocks/hooks/useAssosciatedCompaniesBlock/associated-company-to-workflow-adapter';
 
 const registryInfoWhitelist = ['open_corporates'] as const;
 
@@ -123,6 +124,7 @@ export const useDefaultBlocksLogic = () => {
     mainRepresentative,
     mainContact,
     openCorporate: _openCorporate,
+    associatedCompanies: _associatedCompanies,
     ...entityDataAdditionalInfo
   } = workflow?.context?.entity?.data?.additionalInfo ?? {};
   const { website: websiteBasicRequirement, processingDetails, ...storeInfo } = store ?? {};
@@ -305,7 +307,8 @@ export const useDefaultBlocksLogic = () => {
 
   const ubosRegistryProvidedBlock = useUbosRegistryProvidedBlock(
     ubosRegistryProvided,
-    workflow?.context?.pluginsOutput?.ubo?.message,
+    workflow?.context?.pluginsOutput?.ubo?.message ??
+      workflow?.context?.pluginsOutput?.ubo?.data?.message,
     workflow?.context?.pluginsOutput?.ubo?.isRequestTimedOut,
   );
 
@@ -337,8 +340,16 @@ export const useDefaultBlocksLogic = () => {
     [mutateEvent],
   );
 
+  const associatedCompanies =
+    !kybChildWorkflows?.length &&
+    !workflow?.workflowDefinition?.config?.isAssociatedCompanyKybEnabled
+      ? workflow?.context?.entity?.data?.additionalInfo?.associatedCompanies?.map(
+          associatedCompanyToWorkflowAdapter,
+        )
+      : kybChildWorkflows;
+
   const associatedCompaniesBlock = useAssociatedCompaniesBlock({
-    workflows: kybChildWorkflows,
+    workflows: associatedCompanies,
     onClose,
     isLoadingOnClose: isLoadingEvent,
     dialog: {
@@ -384,7 +395,7 @@ export const useDefaultBlocksLogic = () => {
   });
 
   const associatedCompaniesInformationBlock = useAssociatedCompaniesInformationBlock(
-    kybChildWorkflows ?? [],
+    associatedCompanies ?? [],
   );
 
   const websiteMonitoringBlocks = useWebsiteMonitoringReportBlock();
@@ -400,7 +411,10 @@ export const useDefaultBlocksLogic = () => {
 
   const amlData = useMemo(() => [workflow?.context?.aml], [workflow?.context?.aml]);
 
-  const amlBlock = useAmlBlock(amlData);
+  const amlBlock = useAmlBlock({
+    data: amlData,
+    vendor: workflow?.context?.aml?.vendor ?? '',
+  });
 
   const amlWithContainerBlock = useMemo(() => {
     if (!amlBlock?.length) {