+++ apps/kyb-app/src/get-api-origin/get-api-origin.ts
@@ -1,0 +1,5 @@
+export const getApiOrigin = () => {
+  const url = new URL(import.meta.env.VITE_API_URL);
+
+  return url.origin;
+};