+++ services/workflows-service/src/business-report/business-report.controller.internal.ts
@@ -28,6 +28,7 @@ import { VerifyUnifiedApiSignatureDecorator } from '@/common/decorators/verify-u
 import { BusinessReportHookBodyDto } from '@/business-report/dtos/business-report-hook-body.dto';
 import { BusinessReportHookSearchQueryParamsDto } from '@/business-report/dtos/business-report-hook-search-query-params.dto';
 import { QueryMode } from '@/common/query-filters/query-mode';
+import { isNumber } from 'lodash';
 
 @common.Controller('internal/business-reports')
 @swagger.ApiExcludeController()
@@ -55,6 +56,20 @@ export class BusinessReportControllerInternal {
     }: CreateBusinessReportDto,
     @CurrentProject() currentProjectId: TProjectId,
   ) {
+    const customer = await this.customerService.getByProjectId(currentProjectId);
+
+    const { maxBusinessReports, withQualityControl } = customer.config || {};
+
+    if (isNumber(maxBusinessReports) && maxBusinessReports > 0) {
+      const businessReportsCount = await this.businessReportService.count({}, [currentProjectId]);
+
+      if (businessReportsCount >= maxBusinessReports) {
+        throw new BadRequestException(
+          `You have reached the maximum number of business reports allowed (${maxBusinessReports}).`,
+        );
+      }
+    }
+
     let business: Pick<Business, 'id' | 'correlationId'> | undefined;
     const merchantNameWithDefault = merchantName || 'Not detected';
 
@@ -100,13 +115,19 @@ export class BusinessReportControllerInternal {
     });
 
     const response = await axios.post(
-      `${env.UNIFIED_API_URL}/tld/reports`,
+      `${env.UNIFIED_API_URL}/merchants/analysis`,
       {
         websiteUrl,
         countryCode,
         parentCompanyName: merchantName,
         reportType,
+        withQualityControl,
         callbackUrl: `${env.APP_API_URL}/api/v1/internal/business-reports/hook?businessId=${business.id}&businessReportId=${businessReport.id}`,
+        metadata: {
+          customerId: customer.id,
+          customerName: customer.displayName,
+          workflowRuntimeDataId: null,
+        },
       },
       {
         headers: {