+++ services/workflows-service/src/ui-definition/ui-definition.repository.ts
@@ -33,19 +33,31 @@ export class UiDefinitionRepository {
     );
   }
 
-  async findByWorkflowDefinitionId<
-    T extends Omit<Prisma.UiDefinitionFindFirstOrThrowArgs, 'where'>,
-  >(
+  async findByArgs(args: Prisma.UiDefinitionFindFirstOrThrowArgs, projectIds: TProjectIds) {
+    return await this.prisma.uiDefinition.findFirstOrThrow(
+      this.scopeService.scopeFindFirst(args, projectIds),
+    );
+  }
+
+  async findByWorkflowDefinitionId(
     workflowDefinitionId: string,
     uiContext: keyof typeof UiDefinitionContext,
-    args: Prisma.SelectSubset<T, Omit<Prisma.UiDefinitionFindFirstOrThrowArgs, 'where'>>,
     projectIds: TProjectIds,
+    args?: Prisma.UiDefinitionFindFirstOrThrowArgs,
   ): Promise<UiDefinition> {
     return await this.prisma.uiDefinition.findFirstOrThrow(
       this.scopeService.scopeFindFirst(
         {
-          where: { workflowDefinitionId, uiContext: uiContext },
           ...args,
+          where: {
+            OR: [
+              {
+                workflowDefinitionId,
+                uiContext: uiContext,
+              },
+              ...(args?.where ? [args.where] : []),
+            ],
+          },
         },
         projectIds,
       ),