+++ services/workflows-service/src/ui-definition/ui-definition.service.ts
@@ -1,6 +1,7 @@
-import type { TProjectIds } from '@/types';
+import type { TProjectId, TProjectIds } from '@/types';
 import { UiDefinitionRepository } from '@/ui-definition/ui-definition.repository';
 import { WorkflowRuntimeDataRepository } from '@/workflow/workflow-runtime-data.repository';
+import { replaceNullsWithUndefined } from '@ballerine/common';
 import { Injectable } from '@nestjs/common';
 import { Prisma, UiDefinitionContext } from '@prisma/client';
 
@@ -23,17 +24,21 @@ export class UiDefinitionService {
     return await this.repository.findById(id, args, projectIds);
   }
 
+  async findByArgs(args: Prisma.UiDefinitionFindFirstOrThrowArgs, projectIds: TProjectIds) {
+    return await this.repository.findByArgs(args, projectIds);
+  }
+
   async getByWorkflowDefinitionId(
     workflowDefinitionId: string,
     uiContext: keyof typeof UiDefinitionContext,
     projectIds: TProjectIds,
-    args: Omit<Prisma.UiDefinitionFindFirstOrThrowArgs, 'where'>,
+    args?: Prisma.UiDefinitionFindFirstOrThrowArgs,
   ) {
     return await this.repository.findByWorkflowDefinitionId(
       workflowDefinitionId,
       uiContext,
-      args,
       projectIds,
+      args,
     );
   }
 
@@ -45,12 +50,10 @@ export class UiDefinitionService {
   ) {
     const runtime = await this.workflowRuntimeRepository.findById(runtimeId, {}, projectIds);
 
-    return this.getByWorkflowDefinitionId(
-      runtime.workflowDefinitionId,
-      uiContext,
-      projectIds,
-      args,
-    );
+    return this.getByWorkflowDefinitionId(runtime.workflowDefinitionId, uiContext, projectIds, {
+      ...args,
+      ...(runtime.uiDefinitionId ? { where: { id: runtime.uiDefinitionId } } : {}),
+    });
   }
 
   async list(projectIds: TProjectIds) {
@@ -78,5 +81,19 @@ export class UiDefinitionService {
       },
       projectIds,
     );
+  }
+
+  async cloneUIDefinitionById(id: string, projectId: TProjectId) {
+    const {
+      createdAt,
+      updatedAt,
+      id: _,
+      ...uiDefinition
+    } = await this.repository.findById(id, {}, [projectId]);
+
+    //@ts-ignore
+    const uiDefinitionCopy = await this.create({ data: replaceNullsWithUndefined(uiDefinition) });
+
+    return uiDefinitionCopy;
   }
 }