+++ apps/workflows-dashboard/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ballerine/workflows-dashboard
 
+## 0.2.10
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.19
+  - @ballerine/ui@0.5.12
+
 ## 0.2.9
 
 ### Patch Changes